+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.1.271](https://github.com/fohte/armyknife/compare/v0.1.270...v0.1.271) (2026-08-19)
+
+
+### Features
+
+* **cc:** allow `--worktree` to be omitted ([#831](https://github.com/fohte/armyknife/issues/831)) ([7c76dfb](https://github.com/fohte/armyknife/commit/7c76dfb8b28505dc7c5c88d07155f97e41ba933b))
+
+## [0.1.270](https://github.com/fohte/armyknife/compare/v0.1.269...v0.1.270) (2026-08-19)
+
+
+### Features
+
+* **cc/new:** support choosing the launched session's model via --model ([#829](https://github.com/fohte/armyknife/issues/829)) ([3b12c2b](https://github.com/fohte/armyknife/commit/3b12c2b99ff49f64a050b1a2037eba56a1016d23))
+
 ## [0.1.269](https://github.com/fohte/armyknife/compare/v0.1.268...v0.1.269) (2026-08-19)